# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "This is a sample Laravel API documentation using Swagger",
        "contact": {
            "email": "temuru502@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Login a user and return a token",
                "operationId": "1ff9c65c5cd1a8e0786e318eed856972",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "temuru502@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user",
                "description": "Returns the authenticated user",
                "operationId": "257a57ee6760f86fb4b11ff94dba1f38",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "johndoe@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Logs out the authenticated user",
                "operationId": "eae6a8251a6469bbabe454860f64cf37",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh token",
                "description": "Refreshes the authentication token",
                "operationId": "cbc10c8921f8f6c47d1dd818a3901faf",
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/statistics": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get all statistics",
                "operationId": "5dcc4a1ba1bc86e6f67e79f2f13938f7",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Statistic"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Create a new statistic",
                "operationId": "74953a9b92730f4a4db99434d658d5d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatisticRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Statistic created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Statistic"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/statistics/{id}": {
            "put": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Update an existing statistic",
                "operationId": "88d8bf146b1bd1a317da2846a5f5bc81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Statistic ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StatisticRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statistic updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Statistic"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Delete a statistic",
                "operationId": "d7ebc0ae3a9e1c9e6628c1abffab136d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Statistic ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Statistic deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StatisticRequest": {
                "required": [
                    "unverstet_id",
                    "type_id",
                    "total_students",
                    "total_scholarships"
                ],
                "properties": {
                    "unverstet_id": {
                        "description": "ID of the associated university",
                        "type": "integer",
                        "example": 1
                    },
                    "type_id": {
                        "description": "ID of the type (e.g., program type)",
                        "type": "integer",
                        "example": 1
                    },
                    "total_students": {
                        "description": "Total number of students",
                        "type": "integer",
                        "example": 500
                    },
                    "total_scholarships": {
                        "description": "Total number of scholarships",
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
...
            "Statistic": {
                "required": [
                    "unverstet_id",
                    "type_id",
                    "total_students",
                    "total_scholarships"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "unverstet_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "total_students": {
                        "type": "integer",
                        "example": 500
                    },
                    "total_scholarships": {
                        "type": "integer",
                        "example": 150
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-10T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-10T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
...
    },
...
}